Change the return type of the function "getWeight" of the class IWeightVector to float64.

+++ cpp/subprojects/common/include/common/sampling/weight_vector.hpp
@@ -35,6 +35,6 @@ class IWeightVector {
          * @param pos   The index of the example
          * @return      The weight of the example at the given index
          */
-        virtual uint32 getWeight(uint32 pos) const = 0;
+        virtual float64 getWeight(uint32 pos) const = 0;
 
 };
