+++ cpp/subprojects/common/src/common/sampling/weight_vector_equal.cpp
@@ -14,6 +14,6 @@ bool EqualWeightVector::hasZeroWeights() const {
     return false;
 }
 
-uint32 EqualWeightVector::getWeight(uint32 pos) const {
+float64 EqualWeightVector::getWeight(uint32 pos) const {
     return 1;
 }
